Add request/response pattern to CLJS implementation
The unit tests are a bit tricky, since js is single threaded. There is
a use of timeouts to allow for things to wait, but then the single
threaded nature of js, combined with callbacks, come back to bite us.

The answer is to try and stagger the continuation of the execution of
the thread, allowing for time for the callbacks to execture. Mostly
successful from lots of manual tests, but it's still a bit dodgy at
times and shows non-deterministic behaviour in the tests, with a 90+%
success rate of the tests.